Completed Successful communication between pc and microcontroller.
I've been working on this for a while, and finally I got my communication interface working to the point where it's actually getting useful.

The way it works is that you first send a 16 byte message header that includes all the messaging information for the actual message(which can range in size), then the actual message itself.